chore: Renamed media-query module
